+++ invesdwin-util/src/main/java/de/invesdwin/util/math/decimal/scaled/IPercentData.java
@@ -1,9 +1,7 @@
 package de.invesdwin.util.math.decimal.scaled;
 
-import de.invesdwin.util.math.decimal.Decimal;
-
 public interface IPercentData {
 
-    Decimal getRate();
+    double getRate();
 
 }
